remove the outdated articles
