+++ .openpublishing.redirection.json
@@ -3094,6 +3094,14 @@
             "redirect_url": "/dotnet/csharp/language-reference/operators/namespace-alias-qualifier"
         },
         {
+            "source_path": "docs/csharp/programming-guide/namespaces/how-to-use-the-my-namespace.md",
+            "redirect_url": "/dotnet/csharp/fundamentals/types/namespaces"
+        },
+        {
+            "source_path": "docs/csharp/programming-guide/namespaces/using-namespaces.md",
+            "redirect_url": "/dotnet/csharp/fundamentals/types/namespaces"
+        },
+        {
             "source_path": "docs/csharp/programming-guide/nullable-types/boxing-nullable-types.md",
             "redirect_url": "/dotnet/csharp/language-reference/builtin-types/nullable-value-types"
         },
