--- docs/csharp/programming-guide/namespaces/snippets/Program.cs
@@ -1,9 +1,0 @@
-﻿// See https://aka.ms/new-console-template for more information
-// <Snippet31>
-Console.WriteLine("Hello, World!");
-//</Snippet31>
-
-//<Snippet30>
-System.Console.WriteLine("Hello, World!");
-//</Snippet30>
-
