--- docs/csharp/programming-guide/namespaces/snippets/Usings.cs
@@ -1,120 +1,0 @@
-namespace namespaceexamples
-{
-//<Snippet1>
-using System;
-//</Snippet1>
-
-//<Snippet5>
-using generics = System.Collections.Generic;
-
-//<Snippet14>
-using Alias = System;
-//</Snippet14>
-
-//<Snippet15>
-namespace Library
-{
-    public class C : Alias::Exception { }
-}
-//</Snippet15>
-
-namespace AliasExample
-{
-    class TestClass
-    {
-        static void Main()
-        {
-            generics::Dictionary<string, int> dict = new generics::Dictionary<string, int>()
-            {
-                ["A"] = 1,
-                ["B"] = 2,
-                ["C"] = 3
-            };
-
-            foreach (string name in dict.Keys)
-            {
-                System.Console.WriteLine($"{name} {dict[name]}");
-            }
-            // Output:
-            // A 1
-            // B 2
-            // C 3
-        }
-    }
-}
-//</Snippet5>
-
-//<Snippet82>
-namespace SampleNamespace
-{
-    class SampleClass
-    {
-        public void SampleMethod()
-        {
-            System.Console.WriteLine(
-              "SampleMethod inside SampleNamespace");
-        }
-    }
-
-    // Create a nested namespace, and define another class.
-    namespace NestedNamespace
-    {
-        class SampleClass
-        {
-            public void SampleMethod()
-            {
-                System.Console.WriteLine(
-                  "SampleMethod inside NestedNamespace");
-            }
-        }
-    }
-
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            // Displays "SampleMethod inside SampleNamespace."
-            SampleClass outer = new SampleClass();
-            outer.SampleMethod();
-
-            // Displays "SampleMethod inside SampleNamespace."
-            SampleNamespace.SampleClass outer2 = new SampleNamespace.SampleClass();
-            outer2.SampleMethod();
-
-            // Displays "SampleMethod inside NestedNamespace."
-            NestedNamespace.SampleClass inner = new NestedNamespace.SampleClass();
-            inner.SampleMethod();
-        }
-    }
-}
-//</Snippet82>
-
-//<Snippet9>
-namespace N1     // N1
-{
-    class C1      // N1.C1
-    {
-        class C2   // N1.C1.C2
-        {
-        }
-    }
-    namespace N2  // N1.N2
-    {
-        class C2   // N1.N2.C2
-        {
-        }
-    }
-}
-//</Snippet9>
-
-//<Snippet10>
-namespace N1.N2
-{
-    class C3   // N1.N2.C3
-    {
-    }
-}
-//</Snippet10>
-
-
-}
