--- docs/csharp/programming-guide/namespaces/using-namespaces.md
@@ -1,98 +1,0 @@
----
-title: "Using Namespaces - C# Programming Guide"
-description: Learn how to use namespaces in C# programming, such as accessing namespaces, namespace aliases, and using namespaces to control scope.
-ms.date: 09/10/2021
-f1_keywords: 
-  - "cs.names"
-helpviewer_keywords: 
-  - "fully qualified names [C#]"
-  - "namespaces [C#], how to use"
----
-# Using namespaces (C# Programming Guide)
-
-Namespaces are heavily used within C# programs in two ways. Firstly, the .NET classes use namespaces to organize its many classes. Secondly, declaring your own namespaces can help control the scope of class and method names in larger programming projects.
-
-## Accessing namespaces
-
-Most C# applications begin with a section of `using` directives. This section lists the namespaces that the application will be using frequently, and saves the programmer from specifying a fully qualified name every time that a method that is contained within is used.
-
-For example, by including the line:
-
-[!code-csharp[csProgGuide#1](./snippets/Usings.cs#1)]
-
-At the start of a program, the programmer can use the code:
-
-[!code-csharp[csProgGuide#31](./snippets/Program.cs#31)]
-
-Instead of:
-
-[!code-csharp[csProgGuide#30](./snippets/Program.cs#30)]
-
-## Namespace aliases
-
-You can also use the [`using` directive](../../language-reference/keywords/using-directive.md) to create an alias for a namespace. Use the [namespace alias qualifier `::`](../../language-reference/operators/namespace-alias-qualifier.md) to access the members of the aliased namespace. The following example shows how to create and use a namespace alias:
-
-[!code-csharp[csProgGuideNamespaces#5](./snippets/Usings.cs#5)]
-
-## Using namespaces to control scope
-
-The `namespace` keyword is used to declare a scope. The ability to create scopes within your project helps organize code and lets you create globally-unique types. In the following example, a class titled `SampleClass` is defined in two namespaces, one nested inside the other. The [`.` token](../../language-reference/operators/member-access-operators.md#member-access-expression-) is used to differentiate which method gets called.
-
-[!code-csharp[csProgGuideNamespaces#8](./snippets/Usings.cs#82)]
-
-## Fully qualified names
-
-Namespaces and types have unique titles described by fully qualified names that indicate a logical hierarchy. For example, the statement `A.B` implies that `A` is the name of the namespace or type, and `B` is nested inside it.
-
-In the following example, there are nested classes and namespaces. The fully qualified name is indicated as a comment following each entity.
-
-[!code-csharp[csProgGuideNamespaces#9](./snippets/Usings.cs#9)]
-
-In the previous code segment:
-
-- The namespace `N1` is a member of the global namespace. Its fully qualified name is `N1`.
-- The namespace `N2` is a member of `N1`. Its fully qualified name is `N1.N2`.
-- The class `C1` is a member of `N1`. Its fully qualified name is `N1.C1`.
-- The class name `C2` is used two times in this code. However, the fully qualified names are unique. The first instance of `C2` is declared inside `C1`; therefore, its fully qualified name is: `N1.C1.C2`. The second instance of `C2` is declared inside a namespace `N2`; therefore, its fully qualified name is `N1.N2.C2`.
-
-Using the previous code segment, you can add a new class member, `C3`, to the namespace `N1.N2` as follows:
-
-[!code-csharp[csProgGuideNamespaces#10](./snippets/Usings.cs#10)]
-
-In general, use the [namespace alias qualifier `::`](../../language-reference/operators/namespace-alias-qualifier.md) to reference a namespace alias or `global::` to reference the global namespace and `.` to qualify types or members.
-
-It is an error to use `::` with an alias that references a type instead of a namespace. For example:
-
-```csharp
-using Alias = System.Console;
-
-// Error
-//Alias::WriteLine("Hi");
-
-// OK
-Alias.WriteLine("Hi");
-```
-
-Remember that the word `global` is not a predefined alias; therefore, `global.X` does not have any special meaning. It acquires a special meaning only when it is used with `::`.
-
-Compiler warning CS0440 is generated if you define an alias named global because `global::` always references the global namespace and not an alias. For example, the following line generates the warning:
-
-```csharp
-using global = System.Collections;   // Warning
-```
-
-Using `::` with aliases is a good idea and protects against the unexpected introduction of additional types. For example, consider this example:
-
-[!code-csharp[csProgGuideNamespaces#14](./snippets/Usings.cs#14)]
-
-[!code-csharp[csProgGuideNamespaces#15](./snippets/Usings.cs#15)]
-
-This works, but if a type named `Alias` were to subsequently be introduced, `Alias.` would bind to that type instead. Using `Alias::Exception` ensures that `Alias` is treated as a namespace alias and not mistaken for a type.
-
-## See also
-
-- [C# Programming Guide](../index.md)
-- [Namespaces](../../fundamentals/types/namespaces.md)
-- [Member access expression](../../language-reference/operators/member-access-operators.md#member-access-expression-)
-- [:: operator](../../language-reference/operators/namespace-alias-qualifier.md)
-- [extern alias](../../language-reference/keywords/extern-alias.md)
