+++ docs/csharp/toc.yml
@@ -816,12 +816,6 @@ items:
       href: programming-guide/generics/generics-and-reflection.md
     - name: Generics and Attributes
       href: programming-guide/generics/generics-and-attributes.md
-  - name: Namespaces
-    items:
-    - name: Using namespaces
-      href: programming-guide/namespaces/using-namespaces.md
-    - name: "How to use the My namespace"
-      href: programming-guide/namespaces/how-to-use-the-my-namespace.md
   - name: File System and the Registry
     items:
     - name: Overview
