--- samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideNamespaces/CS/Namespaces.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFramework>net48</TargetFramework>
-    <RootNamespace>CsCsrefProgrammingNamespaces</RootNamespace>
-    <AssemblyName>CsCsrefProgrammingNamespaces</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Reference Include="Microsoft.VisualBasic" />
-  </ItemGroup>
-
-</Project>
